Add navigation drawer and separate import activity
