+++ app/src/main/res/menu/main_drawer.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/nav_import"
+        android:title="@string/nav_import" />
+    <item
+        android:id="@+id/nav_show"
+        android:enabled="false"
+        android:title="@string/nav_show" />
+    <item
+        android:id="@+id/nav_export"
+        android:enabled="false"
+        android:title="@string/nav_export" />
+</menu>
