+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,10 @@
 <resources>
     <string name="app_name">Kindler</string>
+    <string name="app_title">My Highlights</string>
+    <string name="main_placeholder">Select an option from the menu to get started.</string>
+    <string name="navigation_drawer_open">Open navigation drawer</string>
+    <string name="navigation_drawer_close">Close navigation drawer</string>
+    <string name="nav_import">Import</string>
+    <string name="nav_show">Show</string>
+    <string name="nav_export">Export</string>
 </resources>
